Fix bug in rsrc_load() if many objects in rsc
If there were more than 1366 objects in a resource, the calculation
of the offset of objects 1367 upwards was incorrect.  As a result,
random memory overwrites occurred during rsrc_load(), and the objects
were not displayable.  This affected PixArt4.

+++ aes/gemrslib.c
@@ -167,7 +167,7 @@ static RSCITEM get_sub(WORD rsindex, WORD rtype, WORD rsize)
     offset = rs_hdr.wordptr[rtype];
 
     /* get base of objects and then index in */
-    return (RSCITEM)(rs_hdr.base + offset + rsize * rsindex);
+    return (RSCITEM)(rs_hdr.base + offset + (LONG)rsize * rsindex);
 }
